requirements.txt
great_expectations
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.2
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
appnope==0.1.4
argcomplete==3.5.3
asgiref==3.8.1
asttokens==2.4.1
async-generator==1.10
async-timeout==5.0.1
attrs==25.1.0
auto-metamask==0.1.3
babel==2.17.0
blinker==1.9.0
cachelib==0.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.1.1
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2025.2.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
greenlet==2.0.2
grpcio==1.65.5
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy-object-proxy==1.10.0
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.0.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
outcome==1.2.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pip==25.0.1
platformdirs==4.3.6
playwright==1.34.0
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.6
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyee==9.0.4
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
pytest==7.3.1
pytest-base-url==2.0.0
pytest-fixtures==0.1.0
pytest-playwright==0.3.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
robotframework==6.1
robotframework-pythonlibcore==4.1.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.1.0
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.9.1
selenium-stealth==1.0.6
setproctitle==1.3.4
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==3.8.6
Werkzeug==2.2.3
wheel==0.45.1
wirerope==1.0.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
